repair bug with removin no existing file

+++ communication/uart/frdm_connect.py
@@ -35,7 +35,8 @@ def refresh_response():
     return resp
 
 def refresh():
-    os.remove('/tmp/django.lck')
+    if os.path.isfile('/tmp/django.lck'):
+        os.remove('/tmp/django.lck')
     file_object = open(filename, 'w')
     file_object.write(chr(128)+'\n')              #0b10000000 as refresh Question
     file_object.close()
